requirements.txt
absl-py==0.9.0
astor==0.8.1
astunparse==1.6.3
attrs==24.2.0
audioread==3.0.1
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
cffi==1.17.0
chardet==3.0.4
colorama==0.4.6
contourpy==1.1.1
customtkinter==5.2.2
cycler==0.10.0
darkdetect==0.8.0
decorator==4.4.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.6.1
gast==0.3.3
google-auth==1.11.3
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.27.2
h5py==3.11.0
idna==2.9
importlib_metadata==8.2.0
importlib_resources==6.4.3
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
jax==0.4.13
jedi==0.16.0
Jinja2==3.1.4
joblib==1.4.2
jupyter-client==6.1.0
jupyter-core==4.6.3
kapre==0.3.4
keras==2.8.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.2.post1
llvmlite==0.41.1
Markdown==3.2.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mediapipe==0.10.11
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.8
msvc-runtime==14.29.30133
networkx==3.1
numba==0.58.1
numpy @ file:///C:/Users/dev-admin/mkl/numpy_and_numpy_base_1682982345978/work
oauthlib==3.1.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt-einsum==3.2.0
packaging==24.0
pandas==2.0.3
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==4.2.2
pooch==1.8.2
prompt-toolkit==3.0.4
protobuf==3.20.3
psutil==6.0.0
ptyprocess==0.6.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycparser==2.19
pydub==0.25.1
Pygments==2.6.1
pyparsing==2.4.6
python-dateutil==2.9.0.post0
python-speech-features==0.6
pytz==2024.1
pywin32==306
PyYAML==5.3.1
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==1.3.2
scipy==1.9.3
seaborn==0.13.2
six==1.14.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.3.7
sympy==1.13.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.7.0
tensorflow==2.8.0
tensorflow-estimator==2.3.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.5.0
torch==2.4.0
torchvision==0.19.0
tornado==6.0.4
tqdm==4.66.5
traitlets==4.3.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.78
ultralytics-thop==2.0.3
urllib3==1.25.8
wavio==0.0.4
wcwidth==0.1.8
Werkzeug==1.0.0
wincertstore==0.2
wrapt==1.12.0
zipp==3.20.0
